finstock/requirements.txt
# Backend API
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0

# Data & ML
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.2
scikit-learn==1.3.2

# Market data
vnstock
vnquant
yfinance
